fix: resolve build failures and serialization errors
- Move pages/types.ts to types/pages.ts (fix Next.js pages directory issue)
- Fix undefined author serialization in talks, experiments, and workshops
- Add fallback author handling for talks and workshops pages
- Filter out undefined authors with .filter(Boolean) to prevent JSON serialization errors
- Fix flaky shuffle test by using larger array to eliminate chance of same order

Resolves Next.js build optimization failures and undefined property access errors.
